Game module restucture
